Update
> Rollback desktop wallet

+++ SeguraChain-Desktop-Wallet/InternalForm/Startup/ClassWalletStartupInternalForm.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using SeguraChain_Desktop_Wallet.Common;
@@ -13,7 +12,7 @@ namespace SeguraChain_Desktop_Wallet.InternalForm.Startup
 {
     public partial class ClassWalletStartupInternalForm : Form
     {
-        private bool _mainInterfaceStarted;
+        private bool _canOpenMainInterface;
         private bool _noWalletFile;
         private ClassWalletMainInterfaceForm _walletMainInterfaceForm;
 
@@ -35,8 +34,7 @@ namespace SeguraChain_Desktop_Wallet.InternalForm.Startup
         /// <param name="e"></param>
         private void ClassWalletStartupInternalForm_Load(object sender, EventArgs e)
         {
-            if (!_mainInterfaceStarted)
-                OnStartupDesktopWallet();
+            OnStartupDesktopWallet();
         }
 
         #region On startup.
@@ -62,12 +60,10 @@ namespace SeguraChain_Desktop_Wallet.InternalForm.Startup
 
                     _noWalletFile = ClassDesktopWalletCommonData.WalletDatabase.GetCountWalletFile == 0;
 
-                    _walletMainInterfaceForm = new ClassWalletMainInterfaceForm(_noWalletFile, this);
-                    MethodInvoker invoke = () => _walletMainInterfaceForm.ShowDialog(this);
-                    BeginInvoke(invoke);
+                    _canOpenMainInterface = true;
+
                 }
             }).ConfigureAwait(false);
-
         }
 
         #endregion
@@ -77,34 +73,42 @@ namespace SeguraChain_Desktop_Wallet.InternalForm.Startup
         /// <summary>
         /// Function executed on the close of the desktop wallet, close desktop wallet systems.
         /// </summary>
-        public void OnCloseDesktopWallet()
+        public void OnCloseDesktopWallet(ClassWalletMainInterfaceForm walletMainInterfaceForm)
         {
-            Show();
-            _walletMainInterfaceForm.Hide();
+            walletMainInterfaceForm.Hide();
             _walletMainInterfaceForm.Dispose();
 
-            ClassWalletStartupFormLanguage walletStartupFormLanguageObject = ClassDesktopWalletCommonData.LanguageDatabase.GetLanguageContentObject<ClassWalletStartupFormLanguage>(ClassLanguageEnumType.LANGUAGE_TYPE_STARTUP_FORM);
-
-            Text = BlockchainSetting.CoinName + walletStartupFormLanguageObject.FORM_TITLE_CLOSING_TEXT;
-
-            UpdateLabelStartupNoInvokeText(walletStartupFormLanguageObject.LABEL_ON_CLOSE_DESKTOP_WALLET_PENDING_TEXT);
-
-            Refresh();
-
-            Task.Run(async () =>
+            Task.Factory.StartNew(async () =>
             {
+                ClassWalletStartupFormLanguage walletStartupFormLanguageObject = ClassDesktopWalletCommonData.LanguageDatabase.GetLanguageContentObject<ClassWalletStartupFormLanguage>(ClassLanguageEnumType.LANGUAGE_TYPE_STARTUP_FORM);
+                MethodInvoker invoke = Show;
+                BeginInvoke(invoke);
 
-                await ClassDesktopWalletCommonData.CloseDesktopWalletCommonData();
+                invoke = () => Text = BlockchainSetting.CoinName + walletStartupFormLanguageObject.FORM_TITLE_CLOSING_TEXT;
+                BeginInvoke(invoke);
 
-                
-                UpdateLabelStartupText(walletStartupFormLanguageObject.LABEL_ON_CLOSE_DESKTOP_WALLET_SUCCESS_TEXT);
+                UpdateLabelStartupText(walletStartupFormLanguageObject.LABEL_ON_CLOSE_DESKTOP_WALLET_PENDING_TEXT);
 
-                
+                bool onCloseResult = await ClassDesktopWalletCommonData.CloseDesktopWalletCommonData();
+
+                if (onCloseResult)
+                    UpdateLabelStartupText(walletStartupFormLanguageObject.LABEL_ON_CLOSE_DESKTOP_WALLET_SUCCESS_TEXT);
+                else
+                    UpdateLabelStartupText(walletStartupFormLanguageObject.LABEL_ON_CLOSE_DESKTOP_WALLET_FAILED_TEXT);
+
+
                 await Task.Delay(1000);
 
-                Process.GetCurrentProcess().Kill();
-            }).Wait();
+                if (onCloseResult)
+                {
+                    MethodInvoker invokeClose = Close;
+                    BeginInvoke(invokeClose);
+                }
+                // Force to close.
+                else
+                    Process.GetCurrentProcess().Kill();
 
+            }).ConfigureAwait(false);
         }
 
         #endregion
@@ -115,15 +119,19 @@ namespace SeguraChain_Desktop_Wallet.InternalForm.Startup
             {
                 labelStartupDesktopWalletLoadingText.Text = text;
                 labelStartupDesktopWalletLoadingText = ClassGraphicsUtility.AutoSetLocationAndResizeControl<Label>(labelStartupDesktopWalletLoadingText, this, 50d, false);
-                Refresh();
             };
             BeginInvoke(invoke);
         }
 
-        private void UpdateLabelStartupNoInvokeText(string text)
+        private void timerOpenMainInterface_Tick(object sender, EventArgs e)
         {
-            labelStartupDesktopWalletLoadingText.Text = text;
-            labelStartupDesktopWalletLoadingText = ClassGraphicsUtility.AutoSetLocationAndResizeControl<Label>(labelStartupDesktopWalletLoadingText, this, 50d, false);
+            if (_canOpenMainInterface)
+            {
+                timerOpenMainInterface.Stop();
+                _canOpenMainInterface = false;
+                _walletMainInterfaceForm = new ClassWalletMainInterfaceForm(_noWalletFile, this);
+                _walletMainInterfaceForm.ShowDialog(this);
+            }
         }
     }
 }
